backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
smartapi-python>=1.3.0
pyotp>=2.9.0
boto3>=1.34.0
pandas>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
pytz>=2024.1
